requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
config==0.5.1
dataclasses==0.6
Flask==2.0.2
Flask-Mail==0.9.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
pdf-mail==3.0.0
pdfkit==0.6.1
protobuf==3.19.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
twilio==7.2.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
